Updated models and routing
Should now search properly within the database and output the query data. site.db is too large to upload on here so I will upload a copy of it on Microsoft teams. So far it only has data for the query table (rs, pos, ref, alt) but I'll upload an updated one with population data by the end of this week
Note: for chromosome start and end position, the output looks messy, but still queries the database fine. Currently, rs value and genes can only handle single searches.